make use of new AbstractProjectionManagerTest, cleanup
