+++ tests/Projection/MySqlEventStoreReadModelProjectionTest.php
@@ -22,7 +22,7 @@ use ProophTest\EventStore\Pdo\TestUtil;
 /**
  * @group pdo_mysql
  */
-class MySqlEventStoreReadModelProjectionTest extends PdoEventStoreReadModelProjectionTestCase
+class MySqlEventStoreReadModelProjectionTest extends PdoEventStoreReadModelProjectionTest
 {
     protected function setUp(): void
     {
